requirements.txt
cassandra-driver >= 3.26.*
python-dotenv